requirements.txt
python-telegram-bot==22.5
sqlalchemy==2.0.43
aiosqlite==0.20.0
python-dotenv==1.1.1
pydantic==2.12.0
APScheduler==3.11.0

